Fix infinite song search.
> on playlists with more than 100 songs, gospotdl
searched for songs repeatedly because instead of
using the limit as offset, we used previous offset
as the offset (i.e, zero) so it'd just continuously
get only the first 100 songs.

> I just got to know about it because I recently tried
downloading a playlist wiht 125 songs. Anyways, it's
fixed now.

> Also, instead of checking if a `next` link exists to
check if there are more songs, now we check if the
found songs are less than the total songs in the playlist
or album(A bit more straightforward).